requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
alembic>=1.13.1
psycopg2-binary>=2.9.9
redis>=5.0.1

# Machine Learning
scikit-learn>=1.3.2
xgboost>=2.0.2
pandas>=2.1.4
numpy>=1.26.0
joblib>=1.3.2
scipy>=1.8.0

# Data Processing
elasticsearch>=8.11.0
kafka-python>=2.0.2
celery>=5.3.4

# HTTP Clients
httpx>=0.25.2
aiohttp>=3.9.1
requests>=2.31.0

# File handling
aiofiles>=23.2.0

# Security
cryptography>=41.0.8
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0
sentry-sdk>=1.38.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
typer>=0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Network Security
scapy==2.5.0
pyshark==0.6
netaddr==0.10.1

# Threat Intelligence
yara-python==4.3.1
python-whois==0.8.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# Time & Date
python-dateutil==2.8.2
pytz==2023.3

# Configuration
pyyaml==6.0.1
toml==0.10.2

# API Documentation
fastapi-users==12.1.2
fastapi-pagination==0.12.13

# Additional dependencies for security testing
starlette>=0.40.0
typing-extensions>=4.8.0
annotated-types>=0.6.0
